Rename influxdbjson to influxdb
